requirements.txt
ajaxuploader==0.3.8
cairocffi==0.8.0
CairoSVG==2.1.3
cffi==1.11.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==1.11.6
django-bootstrap3==9.1.0
django-bower==5.2.0
django-datetime-widget==0.9.3
django-filter==1.1.0
django-material==1.1.0
djangorestframework==3.7.3
GDAL==2.2.3
html5lib==1.0.1
numpy==1.14.0
pdfrw==0.4
Pillow==5.0.0
psycopg2==2.7.3.2
pycparser==2.18
Pyphen==0.9.4
pyshp==1.2.12
python-magic==0.4.13
pytz==2017.3
six==1.11.0
tinycss2==0.6.1
WeasyPrint==0.42.2
webencodings==0.5.1
xlwt==1.3.0
